requirements.txt
# Core dependencies
fastapi==0.115.0
uvicorn==0.30.6
streamlit==1.39.0
requests==2.32.3
retrying==1.3.4
openai==1.50.2
python-dotenv==1.0.1
cachetools==5.5.0
plotly==5.24.1
pandas==2.2.3
numpy==1.26.4
pytz==2024.2
pydantic==2.9.2
textblob==0.18.0.post0  # Updated, assuming it's used for sentiment analysis

# Data processing
scrapy==2.11.2
fake-useragent==1.5.1
statsmodels==0.14.3
beautifulsoup4==4.12.3
lxml==5.3.0

# Google APIs
google-api-python-client==2.149.0
google-auth-oauthlib==1.2.1

# Reporting
reportlab==4.2.2
openpyxl==3.1.5

# Database
sqlalchemy==2.0.35
alembic==1.13.3

# Development
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-mock==3.14.0

# Redis and async support
aiomysql==0.1.1
asyncpg==0.28.0
redis==5.0.0
aioredis==2.0.1
# Additional utilities
python-dateutil==2.8.2
httpx==0.27.0

# Additional dependencies
pydantic[email]==2.9.2
pydantic[dotenv]==2.9.2
pydantic[email,dotenv]==2.9.2  # Combined for convenience
psycopg2-binary==2.9.9
SQLAlchemy==2.0.30
asyncpg==0.29.0
redis>=5.0
aioredis>=2.0
websockets>=11